comments and cleanup
