+++ app/actions.py
@@ -18,13 +18,28 @@ from uuid import uuid4
 import time
 import bcrypt
 import arrow
+import functools
 
 bp = Blueprint("actions", __name__, url_prefix="/action")
+
+
+def verification_required(view):
+    """View decorator that checks POST['verification'] == "i know what i am doing" """
+
+    @functools.wraps(view)
+    def wrapped_view(**kwargs):
+        # check for verification string
+        if request.form["verification"] != "i know what i am doing":
+            flasher("Verification failed, please try again.", color="warning")
+            return redirect(url_for("dashboard.admin"))
+
+    return wrapped_view
 
 
 @bp.route("/update_cfg", methods=["POST"])
 @admin_login_required
 def update_cfg():
+    """Update site config, for example application window and event start dates."""
     c = get_db().cursor()
 
     c.execute(
@@ -50,16 +65,20 @@ def update_cfg():
     return redirect(url_for("dashboard.edit_config"))
 
 
+@bp.route("/download_csv/export.csv")
+@admin_login_required
+def download_csv():
+    """Download all applicants as a CSV"""
+    conn = get_db()
+    csv_str = create_csv(conn)
+    return Response(csv_str, mimetype="text/csv")
+
+
 @bp.route("/submit_dump", methods=["POST"])
 @admin_login_required
+@verification_required
 def submit_dump():
-
-    # check for verification string
-    if request.form["verification"] != "i know what i am doing":
-        flasher("Verification failed, please try again.", color="warning")
-        return redirect(url_for("dashboard.admin"))
-
-    # import all applicants
+    """Import applicants from a bulk CSV"""
     conn = get_db()
     cursor = conn.cursor()
     count = 0
@@ -72,15 +91,19 @@ def submit_dump():
     return redirect(url_for("dashboard.admin"))
 
 
+@bp.route("/download_db/registrations.sqlite")
+@admin_login_required
+def download_db():
+    """Download the entire SQLite database"""
+    # TODO: Is there enough of a justification for this big of a target?
+    return send_file(current_app.config["DATABASE"])
+
+
 @bp.route("/votes/purge", methods=["POST"])
 @admin_login_required
+@verification_required
 def purge_votes():
-    # check for verification string
-    if request.form["verification"] != "i know what i am doing":
-        flasher("Verification failed, please try again.", color="warning")
-        return redirect(url_for("dashboard.admin"))
-
-    # import all applicants
+    """Purge all votes from the database"""
     conn = get_db()
     c = conn.cursor()
     c.execute("DELETE FROM Votes")
@@ -90,15 +113,10 @@ def purge_votes():
     return redirect(url_for("dashboard.admin"))
 
 
-@bp.route("/download_db/registrations.sqlite")
-@admin_login_required
-def download_db():
-    return send_file(current_app.config["DATABASE"])
-
-
 @bp.route("/vote/submit", methods=["POST"])
 @admin_login_required
 def submit_vote():
+    """Submit a vote"""
     if not "rating" in request.form:
         flasher("Please select a rating!", color="danger")
         return redirect(
@@ -161,17 +179,10 @@ def submit_vote():
         return redirect(url_for("dashboard.applicant", user_id=request.form["user_id"]))
 
 
-@bp.route("/download_csv/export.csv")
-@admin_login_required
-def download_csv():
-    conn = get_db()
-    csv_str = create_csv(conn)
-    return Response(csv_str, mimetype="text/csv")
-
-
 @bp.route("/toggle_hiding")
 @admin_login_required
 def toggle_hiding():
+    """Toggle name redaction for the current user."""
     if not "redacted" in session:
         session["redacted"] = True
     else:
@@ -180,174 +191,10 @@ def toggle_hiding():
     return redirect(url_for("dashboard.rate_queue"))
 
 
-# service actions
-
-
-@bp.route("/service/create", methods=["POST"])
-@admin_login_required
-def service_create():
-    c = get_db().cursor()
-
-    api_key = "".join(str(uuid4()).split("-"))
-    api_secret = "".join((str(uuid4()) + str(uuid4())).split("-"))
-    api_secret_crypt = bcrypt.hashpw(
-        api_secret.encode("ascii"), bcrypt.gensalt())
-
-    c.execute(
-        """
-        INSERT INTO Services (
-            display_name,
-            api_key,
-            api_secret,
-            author_email,
-            active,
-            created
-        ) VALUES (?,?,?,?,1,?)
-    """,
-        [
-            request.form["display_name"],
-            api_key,
-            api_secret_crypt,
-            session["email"],
-            time.time(),
-        ],
-    )
-
-    c.connection.commit()
-
-    flasher(
-        f"Service <code>{request.form['display_name']}</code> added successfully.<br/>API secret: <code>{api_secret}</code>. Copy it down, it will not show again!",
-        color="success",
-    )
-
-    return redirect(url_for("dashboard.list_services"))
-
-
-@bp.route("/service/toggle/<api_key>")
-@admin_login_required
-def service_toggle(api_key):
-    c = get_db().cursor()
-
-    c.execute(
-        """
-        SELECT * FROM Services WHERE api_key=?
-    """,
-        (api_key,),
-    )
-
-    svc = c.fetchone()
-    if svc is None:
-        flasher(f"No such service <code>{api_key}</code>", color="warning")
-        return redirect(url_for("dashboard.list_services"))
-
-    c.execute(
-        """
-        UPDATE Services
-            SET active=?
-            WHERE api_key=?
-    """,
-        [1 if svc["active"] == 0 else 0, api_key],
-    )
-
-    c.connection.commit()
-
-    flasher(f"Service {svc['display_name']} toggled.", color="success")
-
-    return redirect(url_for("dashboard.list_services"))
-
-
-@bp.route("/service/recreate_key", methods=["POST"])
-@admin_login_required
-def service_recreate_key():
-    # check for verification string
-    if request.form["verification"] != "i know what i am doing":
-        flasher("Verification failed, please try again.", color="warning")
-        return redirect(url_for("dashboard.list_services"))
-
-    c = get_db().cursor()
-
-    c.execute(
-        """
-        SELECT * FROM Services WHERE api_key=?
-    """,
-        (request.form["api_key"],),
-    )
-
-    svc = c.fetchone()
-    if svc is None:
-        flasher(
-            f"No such service <code>{request.form['api_key']}</code>", color="warning"
-        )
-        return redirect(url_for("dashboard.list_services"))
-
-    api_secret = "".join((str(uuid4()) + str(uuid4())).split("-"))
-    api_secret_crypt = bcrypt.hashpw(
-        api_secret.encode("ascii"), bcrypt.gensalt())
-
-    c.execute(
-        """
-        UPDATE Services
-            SET api_secret=?
-            WHERE api_key=?
-    """,
-        [api_secret_crypt, request.form["api_key"]],
-    )
-
-    c.connection.commit()
-
-    flasher(
-        f"Service <code>{svc['display_name']}</code> re-keyed successfully.<br/>API secret: <code>{api_secret}</code>. Copy it down, it will not show again!",
-        color="success",
-    )
-
-    return redirect(url_for("dashboard.list_services"))
-
-
-@bp.route("/services/delete", methods=["POST"])
-@admin_login_required
-def service_delete():
-    # check for verification string
-    if request.form["verification"] != "i know what i am doing":
-        flasher("Verification failed, please try again.", color="warning")
-        return redirect(url_for("dashboard.list_services"))
-
-    c = get_db().cursor()
-
-    c.execute(
-        """
-        SELECT * FROM Services WHERE api_key=?
-    """,
-        (request.form["api_key"],),
-    )
-
-    svc = c.fetchone()
-    if svc is None:
-        flasher(
-            f"No such service <code>{request.form['api_key']}</code>", color="warning"
-        )
-        return redirect(url_for("dashboard.list_services"))
-
-    c.execute(
-        """
-        DELETE FROM Services WHERE api_key=?
-    """,
-        (request.form["api_key"],),
-    )
-
-    c.connection.commit()
-
-    flasher(
-        f"Service <code>{svc['display_name']}</code> deleted successfully.",
-        color="success",
-    )
-
-    return redirect(url_for("dashboard.list_services"))
-
-
 @bp.route("/invites/new", methods=["POST"])
 @admin_login_required
 def invite_create():
-
+    """Add an invite for a given user."""
     link = request.form["link"] if "link" in request.form else None
     code = request.form["code"] if "code" in request.form else None
 
@@ -370,6 +217,7 @@ def invite_create():
 @bp.route("/invites/delete/<inv_id>")
 @admin_login_required
 def invite_delete(inv_id):
+    """Delete a given invite"""
     c = get_db().cursor()
     c.execute(
         """
