+++ app/auth.py
@@ -1,5 +1,3 @@
-# -- this is a disaster. --
-
 from flask import (
     Flask,
     redirect,
@@ -11,6 +9,7 @@ from flask import (
 )
 from flask_dance.contrib.google import make_google_blueprint, google
 from flask_dance.contrib.github import make_github_blueprint, github
+from .mymlh import make_mymlh_blueprint, mymlh
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 from oauthlib.oauth2.rfc6749.errors import TokenExpiredError
 import functools
@@ -25,94 +24,36 @@ try:
 except ImportError:
     from flask import _request_ctx_stack as stack
 
-bp = Blueprint("auth", __name__)
-
-google_bp = None
-mlh_bp = None
-github_bp = None
-mymlh = None
+bp = Blueprint("auth", __name__, url_prefix="/auth")
 
 
-def register_google_bp(app):
-    global google_bp  # NOT GOOD
-    google_bp = make_google_blueprint(
+def register_auth_bps(app):
+    """Register all Oauth blueprints"""
+
+    # Google
+    app.register_blueprint(make_google_blueprint(
         client_id=app.config["GOOGLE_CLIENT_ID"],
         client_secret=app.config["GOOGLE_CLIENT_SECRET"],
         scope=["profile", "email"],
         redirect_to="auth.profile",
-    )
-    app.register_blueprint(google_bp, url_prefix="/oauth/admin")
+    ), url_prefix="/oauth/admin")
 
-
-def register_mlh_bp(app):
-    global mlh_bp
-    global mymlh
-
-    mlh_bp = OAuth2ConsumerBlueprint(
-        "mymlh",
-        __name__,
+    # MLH
+    app.register_blueprint(make_mymlh_blueprint(
         client_id=app.config["MLH_CLIENT_ID"],
         client_secret=app.config["MLH_CLIENT_SECRET"],
-        base_url="https://my.mlh.io",
-        token_url="https://my.mlh.io/oauth/token",
-        scope=["email", "education", "demographics"],
-        authorization_url="https://my.mlh.io/oauth/authorize",
-        redirect_to="auth.hacker_profile",
-    )
+        scope=[
+            "email", "education", "demographics"],
+        redirect_to="auth.hacker_profile"
+    ), url_prefix="/oauth/hacker")
 
-    @mlh_bp.before_app_request
-    def set_applocal_session():
-        ctx = stack.top
-        ctx.mymlh_oauth = mlh_bp.session
-
-    mymlh = LocalProxy(functools.partial(_lookup_app_object, "mymlh_oauth"))
-    app.register_blueprint(mlh_bp, url_prefix="/oauth/hacker")
-
-
-def register_github_bp(app):
-    global github_bp
-    github_bp = make_github_blueprint(
+    # GitHub
+    app.register_blueprint(make_github_blueprint(
         client_id=app.config["GITHUB_CLIENT_ID"],
         client_secret=app.config["GITHUB_CLIENT_SECRET"],
         scope="user:email",
-        redirect_to="auth.hacker_profile_gh",
-    )
-
-    app.register_blueprint(github_bp, url_prefix="/oauth/hacker")
-
-
-@bp.route("/")
-def index():
-    if google.authorized:
-        return redirect(url_for("auth.profile"))
-    return render_template("auth/login.html")
-
-
-@bp.route("/logout")
-def logout():
-    # retrieve token
-
-    try:
-        token = google_bp.token["access_token"]
-
-        # revoke permission from Google's API
-        resp = google.post(
-            "https://accounts.google.com/o/oauth2/revoke",
-            params={"token": token},
-            headers={"Content-Type": "application/x-www-form-urlencoded"},
-        )
-        assert resp.ok, resp.text
-
-    except TokenExpiredError as e:
-        print("token expired, ignoring")
-    except TypeError as e:
-        print("probably fine too")
-
-    try:
-        del google_bp.token  # Delete OAuth token from storage
-    except Exception as e:
-        print("probably also fine")
-    return redirect(url_for("landing.index"))
+        redirect_to="auth.hacker_profile",
+    ), url_prefix="/oauth/hacker")
 
 
 def admin_login_required(view):
@@ -132,6 +73,7 @@ def hacker_login_required(view):
 
     @functools.wraps(view)
     def wrapped_view(**kwargs):
+        print(mymlh.authorized)
         if not mymlh.authorized and not github.authorized:
             return redirect(url_for("landing.index"))
         return view(**kwargs)
@@ -139,26 +81,60 @@ def hacker_login_required(view):
     return wrapped_view
 
 
+@bp.route("/")
+def index():
+    if google.authorized:
+        return redirect(url_for("auth.profile"))
+    return render_template("auth/login.html")
+
+
+@bp.route("/logout")
+def logout():
+    """Log out an admin (google) account, invalidating the oauth2 token"""
+
+    try:
+        token = google.token["access_token"]
+
+        # revoke permission from Google's API
+        resp = google.post(
+            "https://accounts.google.com/o/oauth2/revoke",
+            params={"token": token},
+            headers={"Content-Type": "application/x-www-form-urlencoded"},
+        )
+        assert resp.ok, resp.text
+
+    except TokenExpiredError as e:
+        pass  # Token expired, No problem
+    except TypeError as e:
+        pass  # Apparently also fine
+
+    try:
+        # Delete OAuth token from storage
+        del google.token
+    except Exception as e:
+        pass  # No token to delete
+
+    return redirect(url_for("landing.index"))
+
+
 @bp.route("/profile")
 @admin_login_required
 def profile():
-    # hacky, but this pins the google profile information to the session
+    # Pin google profile information to the session
     resp = requests.get(
         "https://people.googleapis.com/v1/people/me",
         params={"personFields": "names,emailAddresses,photos"},
         headers={"Authorization": "Bearer {}".format(
-            google.token[u"access_token"])},
+            google.token["access_token"])},
     )
 
     person_info = resp.json()
-    pprint(person_info)
 
     profile = {
-        "email": person_info[u"emailAddresses"][0][u"value"],
-        "name": person_info[u"names"][0][u"displayName"],
-        "image": person_info[u"photos"][0][u"url"].split("=")[
-            0
-        ],  # remove the 100px limit (ends with =s100)
+        "email": person_info["emailAddresses"][0]["value"],
+        "name": person_info["names"][0]["displayName"],
+        # Remove 100 pixel parameter
+        "image": person_info["photos"][0]["url"].replace("=s100", "")
     }
 
     session["name"] = profile["name"]
@@ -170,40 +146,9 @@ def profile():
 
 # -- MLH profile info --
 
-
-@bp.route("/hacker")
-@hacker_login_required
-def hacker_profile():
-    session["login_type"] = "mlh"
-    profile = get_hacker_mlh_profile()
-    session["email"] = profile["email"]
-    session["mlh_info"] = profile
-    return redirect(url_for("hacker.init_mlh"))
-
-
-@bp.route("/hacker/logout")
-@hacker_login_required
-def hacker_logout():
-    try:
-        del mlh_bp.token
-    except:
-        pass
-    try:
-        del github_bp.token
-    except:
-        pass
-
-    session.clear()
-    return redirect(url_for("landing.index"))
-
-
 def get_hacker_mlh_profile():
-    resp = mlh_bp.session.get("/api/v3/user.json")
-    print(resp.json())
+    resp = mymlh.session.get("/api/v3/user.json")
     return resp.json()["data"]
-
-
-# -- GH profile info --
 
 
 def get_hacker_gh_profile():
@@ -217,12 +162,39 @@ def get_hacker_gh_profile():
     return resp
 
 
-@bp.route("/hacker/github")
+@bp.route("/hacker")
 @hacker_login_required
-def hacker_profile_gh():
-    session["login_type"] = "github"
-    profile = get_hacker_gh_profile()
-    session["gh_info"] = profile
-    session["email"] = profile["email"]
+def hacker_profile():
+    mlh = mymlh
 
-    return redirect(url_for("hacker.init_gh"))
+    if mymlh.authorized:
+        # MLH
+        session["login_type"] = "mlh"
+        profile = get_hacker_mlh_profile()
+        session["email"] = profile["email"]
+        session["mlh_info"] = profile
+        return redirect(url_for("hacker.init_mlh"))
+    else:
+        # GitHub
+        session["login_type"] = "github"
+        profile = get_hacker_gh_profile()
+        session["gh_info"] = profile
+        session["email"] = profile["email"]
+
+        return redirect(url_for("hacker.init_gh"))
+
+
+@bp.route("/hacker/logout")
+@hacker_login_required
+def hacker_logout():
+    try:
+        del mymlh.token
+    except:
+        pass
+    try:
+        del current_app.blueprints['github'].token
+    except:
+        pass
+
+    session.clear()
+    return redirect(url_for("landing.index"))
