+++ app/templates/dashboard/base.html
@@ -72,7 +72,7 @@
                         Logout
                     </a>
                     <hr class="navbar-divider">
-                    <a class="navbar-item" href="mailto:patrick@comp-soc.com">
+                    <a class="navbar-item" href="mailto:oscar@hacktheburgh.com">
                         Report an issue
                     </a>
                 </div>
